// old_app_for _References/core/models/shared/actionsConfig.ts
export interface ActionsConfig {
  title?: string;
  showCreate?: boolean;
  showCopy?: boolean;
  showView?: boolean;
  showEdit?: boolean;
  showPost?: boolean;
  showDelete?: boolean;
  showBack?: boolean;
  showSearch?: boolean;
  showRight?: boolean;
  showRefresh?: boolean;
  showexportExcel?: boolean;
  showExportPdf?: boolean;
  showPrint?: boolean;
  showReverse?: boolean;
  showResetPassword?: boolean;
  ShowundoShipment?: boolean;
  showInformation?: boolean;
  showAdvanceFilter?: boolean;
  searchPlaceHolder?: string;
}